requirements.txt
# pip install -r requirements.txt

# Base ----------------------------------------
gitpython
ipython  # interactive notebook
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.7.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.8.1
tqdm>=4.64.0

# Logging -------------------------------------
tensorboard>=2.4.1

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Extras --------------------------------------
-e git+https://github.com/ddcanedo/albumentations.git@e30231cf3dc27ff972075f602e3bbc152c81e154#egg=albumentations
rasterio==1.2.10

scikit-learn==1.1.2
Flask==2.2.2
laspy==2.3.0
jakteristics==0.5.0
Shapely==1.8.5.post1
Pyqtree==1.0.0
